Clean up build system, fix test programs
Clean up various left over commented lines from dlopen to single
libhamlib transition.  Remove unneeded configure variables.

Correct minor inconsistencies in Makefile.am files.

Define a new variable, READLINE_LIBS, so that only programs that offer
Readline support are linked against it.

Fix various compilation warnings and errors in test files revealed with
'make check' on MinGW.

Define rig and rotor backends to (mostly) be built in alphabetical
order.

+++ tests/testfreq.c
@@ -28,11 +28,11 @@ int main (int argc, char *argv[])
 #endif
 
 	printf("%s\n", hamlib_version);
-	printf("caps size: %lu\n", sizeof(struct rig_caps));
-	printf("state size: %lu\n", sizeof(struct rig_state));
-	printf("RIG size: %lu\n", sizeof(struct rig));
-	printf("freq_t size: %lu\n", sizeof(freq_t));
-	printf("shortfreq_t size: %lu\n", sizeof(shortfreq_t));
+	printf("caps size: %lu\n", (long unsigned) sizeof(struct rig_caps));
+	printf("state size: %lu\n", (long unsigned) sizeof(struct rig_state));
+	printf("RIG size: %lu\n", (long unsigned) sizeof(struct rig));
+	printf("freq_t size: %lu\n", (long unsigned) sizeof(freq_t));
+	printf("shortfreq_t size: %lu\n", (long unsigned) sizeof(shortfreq_t));
 
 	/* freq on 31bits test */
 	f = GHz(2);
